+++ HISTORY.rst
@@ -1,6 +1,12 @@
 Release History
 ~~~~~~~~~~~~~~~
 
+0.6.0
+-----
+
+* Add support for pytest-cov uploading coverage results and the htmlcov as an artifact
+* Known issue: requires `--cov-report html` to be added on the CLI for the report upload to work
+
 0.5.0
 -----